[filterobject] Move functions to object processing
